Factor out graph.rs
